Fix EIA Conversion to Int

+++ gridstatus/eia.py
@@ -548,7 +548,7 @@ def _handle_region_data(df):
         },
     )
 
-    df["MW"] = df["MW"].astype("Int64")
+    df["MW"] = df["MW"].astype(float)
 
     # pivot on type
     df = df.pivot_table(
@@ -561,7 +561,7 @@ def _handle_region_data(df):
 
     # fix after pivot
     for col in ["Load", "Net Generation", "Load Forecast", "Total Interchange"]:
-        df[col] = df[col].astype("Int64")
+        df[col] = df[col].astype(float)
 
     return df
 
@@ -609,7 +609,7 @@ def _handle_fuel_type_data(df):
         axis=1,
     )
 
-    df["MW"] = df["MW"].astype("Int64")
+    df["MW"] = df["MW"].astype(float)
 
     # pivot on type
     df = df.pivot_table(
@@ -622,7 +622,7 @@ def _handle_fuel_type_data(df):
 
     # nans after pivot because not
     # all respondents have all fuel types
-    df[fuel_mix_cols] = df[fuel_mix_cols].astype("Int64").fillna(0)
+    df[fuel_mix_cols] = df[fuel_mix_cols].astype(float).fillna(0)
 
     df.columns.name = None
